refactor(mespapiers): Extract CategoryItemByPaper from PaperGroup
and modify List construction to be compliant with cozy-ui
